[take 2] refactor(container-runtime): Stop estimating batch size when accumulating a local batch (#24363)
_Re-do of #24309 with less strict assertion conditions in
`messageSize.spec.ts`. See df114eac27269bd2c7b90c1e8125034c68e96c98_

## PR Description (from #24309)

We are preparing to update `LocalBatchMessage` to hold the original
runtime op, not the serialized op. So it will no longer be practical to
estimate the content size of each op (we don't want to do an extra JSON
stringify just for this purpose).

Here's a comparison of the places we would fail before / after this
change:

Before:
* If no compression, then if est socket size is too big, throw when
submitting. (in `Outbox.addMessageToBatchManager`)
* "If no compression" because `hardLimit` would be infinity if
compression was enabled
* If yes compression (and it happens), then throw if it was insufficient
(in `Outbox.virtualizeBatch`)

Now:
* If yes compression (and it happens), then throw if it was insufficient
(in `Outbox.virtualizeBatch`) -- _unchanged from before_
* Otherwise (no compression happened, or it did and appeared
sufficient), if est socket size is too big, throw before send (in
`Outbox.sendBatch`)
* This case used to be possible even after compression if there were
soooo many empty placeholders that it pushed it over the limit. The log
I'm deleting from there ("LargeBatch") is hit for some internal apps on
old versions, when we did empty placeholders. The log has not been hit
on any recent version.
* But now, if compression is sufficient or unnecessary, this will surely
pass as well. So this is just guarding the case where compression was
disabled, so it's quite equivalent to the original check during submit.