add missing unit tests for tableSchema (#24289)
## Description

Adds missing PR tests to `tableSchema.spec.ts`. Some of the tests added
in this PR are skipped, as there are certain cases where we do not have
a policy for yet (ex: inserting existing row, deleting invalid cell,
etc.). Once we come up with a policy for these cases, we can update the
usage errors and un-skip the tests.

Also adds overloads to the `createTable` method to allow it to be called
without passing in a `Row` or `Column` schema.