Make ids allocated by createSnapshotCompressor actually deterministic (#24365)
## Description

Migrate SharedTree test util "getView" to use "independentView" with
createSnapshotCompressor.

Fix createSnapshotCompressor so it's produced IDs are actually
deterministic (and thus safe to use in snapshot tests).

This required changing createAlwaysFinalizedIdCompressor so the ids it
allocates are based on the session id passed into it.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).