build(client): Generate PACKAGES.md in build (#22011)
Builds the PACKAGES.md file as part of the client build. It's run as
part of the "build"/"ci:build" phase.

The PACKAGES.md file is not generated in build because historically, it
has not generated stable markdown cross-platform, so some people see
local changes. But that also means it gets outdated, and in that case,
is it even useful? For now, I suggest we run it as part of the build,
see if that causes issues again (it's been a couple of years), and if
so, disable it and remove or replace it with something else.

Node: Due to historical oddities, the PACKAGES.md file is generated
using the `check:layers` command. Yes, it's a little strange.