build: Skip redundant Component Governance checks (#24371)
## Description

Removes some redundant Component Governance checks:
- In the build pipeline for the client release group we split building
the repository into one job and running tests in separate jobs. Some of
the latter were running the CG tasks even though the build one already
did, because they check out the repo again. They now won't.
- We [recently introduced a new
job](https://github.com/microsoft/FluidFramework/pull/24284) to create a
tag in the repo after publishing a public release, as a fix for a
previous change. That job is also running the CG checks unnecessarily
since the build already happened at that point, and we only check out
the repo to be able to create a tag in it.