Don't bind handles in getGCData (and rename SummarySerializer to GCHandleVisitor) (#24260)
The so-called "SummarySerializer" isn't used for serialization at all,
it's just used as a way to visit all the handles in a given DDS as part
of `getGCData`.

This PR is mostly renames and comments (`SummarySerializer` to
`GCHandleVisitor`).

But there is one important behavior change: **Handles are no longer
bound during this "serialization" pass**.

This should be a no-op, for a few reasons:

* Binding is only relevant in the presence of local changes (otherwise
it's a no-op anyway). Ideally, there are no local changes in the
Summarizer client. Interactive clients don't run GC so `getGCData` won't
be called.
* If there are local changes in Summarizer, such that this bind step is
not a no-op, it will also be about to happen during Summarize itself (GC
always precedes Summarize). In either case, bind would result in an
attach op, either before Summarization starts or during Summarization,
and in either case the Summarizer notices and tries again, waiting for
the local changes to roundtrip.