Remove direct use of TreeFactory and SharedTree class (#24377)
## Description

The SharedTree class only exists to wrap SharedTreeKernel up in a
ISharedObject for use with Fluid's factories and runtimes.

Thus stop using the class and factory directly so that how this adaption
of the kernel into a SharedObjectKind is done is no longer implicitly
depended on my tests.

This enables the refactor in show this is done from
https://github.com/microsoft/FluidFramework/pull/23729 to not be nearly
as disruptive.