(refactor-dds): Update eslint config in packages/dds/shared-summary-block (#24381)
## Description

The package currently uses our "minimal" config. We update it to use
"recommended" and fix the resulting linter violations.

To update the config, replace the
"@fluidframework/eslint-config-fluid/minimal" entry in the extends block
of the package's `.eslintrc.js` with
"@fluidframework/eslint-config-fluid/recommended". Running `npm run
eslint` from within the package directory will reveal the issues that
need to be resolved.