refactor(fluid-static): Add flexibility to createDOProviderContainerRuntimeFactory() (#24305)
## Description

Updates `createDOProviderContainerRuntimeFactory()` so it can optionally
take in a `IFluidDataStoreRegistry` to use in the root data object
factory, so callers can pass one in instead of always using the default
one. Motivated by a scenario where we want to use a custom registry that
doesn't prevent a container from loading when there are unknown
components/DataStores in the snapshot.