(refactor-undo): Update eslint config in packages/framework/undo-redo to extend "recommended" (#24382)
## Description

The package currently uses our "minimal" config. We should update it to
use "recommended" and fix the resulting linter violations.

To update the config, replace the
"@fluidframework/eslint-config-fluid/minimal" entry in the extends block
with "@fluidframework/eslint-config-fluid/recommended". Re-running the
lint command from within the package directory will reveal the issues
that need to be resolved.