feat(container-runtime): Retain original runtime op in pending local state, as opposed to the serialized op (#24364)
Today we stringify before putting the op into the Outbox and the
PendingStateManager. This means that flows like Resubmit and Rollback
don't get the original op, but rather a lossy string-roundtripped copy.

Instead, we can use the original op throughout the ContainerRuntime, and
wait to stringify until virtualizing the batch of ops (grouping,
compressing, chunking) on the way out of the outbox to the ordering
service.