(tree): KeyedFieldEncoder related renames and doc updates (#24389)
## Description

- In `KeyedFieldEncoder`, renamed the property `shape` to `encoder`. The
type of this property is `FieldEncoder` so `encoder` seems like a more
appropriate name.
Most of the shapes implementations are also encoders so the name `shape`
makes sesne from that perspective. But, that is just an implementation
detail and it is a little confusing to see code such as
`fieldEncoder.shape.encode` where shape is of type `FieldEncoder`.
- In `NodeShape`, renamed the property `fields` to
`specializedFieldEncoders`. The type of this property is
`KeyedFieldEncoders`, so having the term `encoder` in the name seems
more appropriate. Also, it contains the keys of fields for which
specialized encoding is provided for optimization.
- In `NodeShape`, renamed the property `extraLocal` to
`otherFieldsEncoder`. This property is an encoder for fields that aren't
in `specializedFieldEncoders`.
- In `NodeShape`, renamed the property `explicitKeys` to
`specializedFieldKeys`. This property contains the keys of fields which
are encoded using `specializedFieldEncoders`.