ContainerRuntime: Only wrap delta manger if summarizer (#24408)
This change is in preparing for projecting readonly state to the
datastore runtime. This change modifies and simplifies the proxy for the
summarizer by moving the summarizer check to a static function, and only
wrapping when the client is a summarizer. This change also fixes a bug
in the proxy where readonly events would still be fired, but their value
wouldn't match that of the proxy