feat(DDS): Only encode handles in DDS layer if ContainerRuntime won't do it (#24409)
We recently started encoding handles in the ContainerRuntime layer.
However, due to compat requirements for DataStore/Runtime boundary, we
can't _stop_ encoding in the DDS layer yet. So the code in
ContainerRuntime is a no-op and we don't get the benefits intended by
this change.

This PR adds a check of the ContainerRuntime's layer compat details, and
if it's going to encode the handles, we don't do it in the DDS layer.