refactor(client-presence): Rename `StatesWorkspace.props` to `StatesWorkspace.states` and  `NotificationsWorkspace.props` to `NotificationsWorkspace.notifications` (#24419)
## Description
This change renames `StatesWorkspace.props` to `StatesWorkspace.states`
and `NotificationsWorkspace.props` to
`NotificationsWorkspace.notifications`. Since NotificationsWorkspace is
meant to be a subset of StatesWorkspace and the APIs now diverge, I use
the NotificationsWorkspaceAdapter class to reconcile this (takes in a
StatesWorkspace and returns its `states` object when `notificaitons` is
called. We return this adapter when Presence.notifications.getWorkspace
is called.