Runtime: Expose readonly directly off the datastore runtime (#24410)
This change plumbs through the readonly flag from the container runtime
to the data store runtime. This will reduce reliance on the delta
manager in datastores as one of the primary reasons it is used is to
retrieve the readonly flag. In addition to the flag, this change also
propagates the readonly event, so changes can be detected.

In order to be backward compatible, there is code in the
DataStoreContext which ensures that the readonly flag is correctly set
for older data stores. In the FluidDataStoreRuntime this is code to
handle context's which do and don't support setting readonly. These
together ensure a consistent experience for developers, regardless of
the versions they are using.

The primary motivation for this change is to leverage readonly to deal
with back and forward compat between layers, as we can build upon this
to support putting specific datastore into readonly if they do not
support some feature. I plan to leverage this for staging mode, so
datastores which should not submit changes during staging mode, can be
placed in readonly.


[AB#32463](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/32463)

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>
Co-authored-by: Kian Thompson <102998837+kian-thompson@users.noreply.github.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>