Make InsertableTreeNodeFromAllowedTypes not distribute over unions (#24441)
## Description

While looking into the non-determinism in JsonAsTree's incremental
build, I noticed InsertableTreeNodeFromAllowedTypes implicitly
distributes over it's extends clause, which is wrong for input types
with contravariance.

I have thus removed the contravariance. This did not appear to help with
the nondeterministic incremental build issue, but its more correct
typing.

## Breaking Changes

Code which provides a schema which is typed as a union of arrays like
`[A, B] | [A]` to an API which accepts allowed types (like the value
kind for a Map node) could be impacted. See new test and changeset.