Add experimental "Staging Mode" for holding changes locally to be either committed all at once, or discarded (#23783)
This introduces a new "Staging Mode" API to the runtime. When in
Staging Mode, local changes are "Staged" and we don't try to submit them
to the ordering service at all. You can exit staging mode by either
"discarding" or "committing" the staged changes.

**Discard** is similar to rollback, but more complex since remote
changes that have been merged with the local changes need to be
preserved even as the local changes are reverted.

**Commit** simply rebases and submits the local changes.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>
Co-authored-by: Matt Rakow <marakow@microsoft.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>