Re-enable unsummarized ops test with larger timeout (#24461)
This test has been flaky in the test pipeline, and always related to
timeouts. Running the test many times locally would eventually produce
flakiness, but only after running the test almost 2000 times in a row.
This leads me to believe this test is affected by service slowdowns more
than usual, and flakiness is mainly coming from service latency. For
example, after running the test about 3000 times, it starts to time out
consistently.

The default timeout for local service is 2000ms, so I bumped this test
to 5000ms for the time being. If the test still continues to be flaky,
we can revisit if this test is still necessary to keep around.

### Test Data

| Number of runs | Timeout | Failures |
| -------- | ------- | ------- |
| 1000 | 2s | 0 |
| 2000 | 2s | 2 |
| 2500 | 2s | 10 |
| 2500 | 2s | 12 |
| 2500 | **3s** | 3 |
| 2500 | **5s** | 0 |


[AB#29483](https://dev.azure.com/fluidframework/internal/_workitems/edit/29483)