Add deprecation comment to `Loader` class without @deprecation tag (#24453)
## Description

To make the `Loader` class internal, first we need to make sure to
remove all of the uses of the class in external repositories.
This change is not meant to deprecate the loader class, but to make
customers aware that it should no longer be used but instead use the
free-form functions.

See [issue
#24450](https://github.com/microsoft/FluidFramework/issues/24450) for
more context.

## Reviewer Guidance

Please let me know if there's something I should change or be aware of.

Fixes
[AB#36354](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/36354)