Nested DataStores: Remove layer compat until feature is complete (#24465)
The nested datastore feature is on hold, and incomplete. This change
reduces the burden of supporting it by removing the need for cross layer
compat. This will continue to support testing and development of the
feature. This change strikes a better balance between a non-final
feature, and other feature development as support.

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>