improvement(tree): Make "props" parameters optional in Column and Row schema factories (#24507)
Adds factory overloads to allow users to not specify "props" when
creating Column / Row schema. For now, this results in an optional field
of type `null`. In the future, I will explore options for omitting the
"props" field (from the interface) when not specified.

See updated tests for more details.