Adding calling service name in S2S request headers (#24518)
## Description

This PR adds the following support for internal S2S API calls:

1) All requests sent within the server will now have the caller name in
the headers with header name `x-calling-service`.
2) The receiving logging middleware logs the calling service name

## Breaking Changes

This is a breaking change as it introduces a new arg in the
`bindTelemetryContext()` function - `serviceName`. I will be
implementing this change as part of a r11 bump in Historian and also in
AFR.