container-runtime: Add API for default configuration (#24422)
## Description

This PR adds an API that allows customers to set leverage the default
configuration logic added in
https://github.com/microsoft/FluidFramework/pull/24379. The API is an
**optional** property added to the `LoadContainerRuntimeParams`
interface, which is used when instantiating a container runtime.

This PR also renames the API from `compatibilityVersion` ->
`minVersionForCollab` (it was changed from `compatibilityMode` ->
`compatibilityVersion` in
https://github.com/microsoft/FluidFramework/pull/24407).

### Considerations

The API was added to `LoadContainerRuntimeParams` for two reasons:
1. Allow us to give clear guidance to customers. For example, "Pass in
`minVersionForCollab` X to ensure collaboration works with clients
running runtime version X.
2. It provides a simple path to requiring customers to pass in a
`minVersionForCollab` in the future (by making the param required).
 
One alternative is to expose a helper function that will generate the
`runtimeOptions` object that is used when instantiating a container
runtime. However, this may complicate the guidance given to customers
and does not provide a clear path to making this a required step in the
future.

## Example

A customer may use the API like so:
```ts
const containerRuntime = await ContainerRuntime.loadRuntime({
    context,
    registryEntries,
    existing,
    provideEntryPoint,
    runtimeOptions: {
	enableGroupedBatching: false,
    },
    minVersionForCollab: "2.20.0", 
});
```
There are two noteworthy observations:
1. `minVersionForCollab` is set to `"2.20.0"`. This will ensure the
default configurations for `IContainerRuntimeOptionsInternal` are set
such that we will ensure that clients running version 2.20.0 or later of
the runtime can collaborate.
5. We also manually set `enableGroupedBatching` in the `runtimeOptions`
param. Batching is normally enabled by default when
`minVersionForCollab` is set to `"2.0.0"` or later, but any
configurations manually set in `runtimeOptions` will override the
default configurations.

## Next Steps
- Add "fail fast" mechanism that prevents explicity setting
`runtimeOptions` in such a way that contradicts `minVersionForCollab`.
For example, setting `enableGroupedBatching` to be `true` and
`minVersionForCollab` to `"1.4.0"`. This should not be allowed since
clients running v1.4.0 of the runtime will not be able to collaborate if
batching is enabled. See
[AB#37810](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/37810)

## Misc

[AB#36088](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/36088)

---------

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>
Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>