Remove getContainerPackageInfo public api from container.ts (#24525)
With PR: https://github.com/microsoft/FluidFramework/pull/22849 we
introduced a non-generic api into container-loader layer which we should
remove.
Since it is public, we will first mark it as deprecated, and then in the
next release remove it. Make sure we inform Loop to not take dependency
on it.

This is the follow up task to make sure that the api gets eliminated
properly.

Reviewer Guidance
Please let me know if there is anything I should keep in mind while
doing this work

Fixes:
[AB#31203](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31203)