Assert that RemoteFluidObjectHandle can't be a bind source (#24402)
## Background

When a new DataStore or DDS is created, it only become attached once its
handle is stored in a fully-attached DDS. To facilitate this,
`IFluidHandleInternal` includes a `bind` function to be implemented by
the "source handle" - the handle to the DDS referencing the target
object.

`bind` takes in the target handle and ensures that target object is
attached as soon as the source object becomes attached (or right away if
it already is).

So, when would a `RemoteObjectFluidHandle` be the source handle? It
can't be - the handle binding flow always uses the SharedObject's own
handle.

## Why enforce this?

We are working on changes to the handle binding flow implementation, and
this helps clarify the current supported behavior.