CG: Updates serialize-javascript to 6.0.2 (#24383)
This change updates `serialize-javascript` from 6.0.0 to 6.0.2 to
address CVE-2024-11831.

`serialize-javascript` is a transitive dependency that comes from Mocha.
Most uses of `serialize-javascript` were resolved by updating mocha from
10.1.0 to 10.8.2. The only package that couldn't be updated in this way
is `eslint-fluid-config`, whose dependency on `mocha-multi-reporters`
pulls in mocha@10.4.0, which in turn brings in
serialize-javascript@6.0.0. To fix that specific package, we've
introduced an override on serialize-javascript to the package JSON

---------

Co-authored-by: Craig Macomber (Microsoft) <42876482+CraigMacomber@users.noreply.github.com>