improvement(container-runtime): Log minVersionForCollab in ContainerLoadStats telemetry event (#24563)
## Description

This PR logs the `minVersionForCollab` used when loading the runtime in
the `ContainerRuntime:ContainerLoadStats` telemetry event.

This will be more relevant when
[AB#31226](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31226)
is complete. In that work item, we plan to send telemetry events when a
client with a verison less than `minVersionForCollab` is tries load the
session. It will be useful for customers to be able to monitor session's
`minVersionForCollab` going forward.