Added abort controller support for S2S Axios requests (#24548)
## Description

This PR adds an abort controller support for S2S Axios requests. These
signals are used to abort a request in case a client disconnects before
the request is completed. This will prevent further ops on the server
side and fail requests faster. Currently, these abort signals are
supported only for HTTP ingress services (Alfred and TODO in Historian)
that make internal FRS S2S calls.

The implementation leverages handling canceled errors in our
`restWrapper` class. It also leverages middleware functions to add event
listeners to trigger the `abort` process. To make these signals
available, I created a global context that lives during the duration of
the request. This helps use abort signals in the context of a client
request.