Add SharedObjectFromKernel (#24401)
## Description

Provide `makeSharedObjectKind`, a new API for implementing ISharedObject
which does not rely on subclassing.

Note that all the new API and abstractions here are `@internal` and thus
could be changed or removed in the future. Additionally the new logic
builds SharedObjects using the old APIs so it has no compat implications
for the runtime side: this is simply providing a new option for how a
SharedObject like SharedTree can be factored.

Currently the approach used to define a DDS is to subclass SharedObject.

This makes the logic used to implement the DDS not very composable. If
someone wanted to use the logic from two different DDSes to implement a
single one (for example to perform a migration, or to reuse the logic in
multiple places like Directory almost does with map)) it is difficult.

This approach (Subclassing SharedObject) can also make targeted testing
more difficult (for example requiring more mocks), since SharedObject
has a significant amount of coupling to runtime types.

Further more, the existing pattern has more boilerplate than necessary
as the existing pattern of implementing IChannelFactory for each
SharedObject is repetitive.

Additionally, this subclassing based pattern tends to lend itself toward
exporting classes, which lead to the use of both SharedObject and the
factory classes in exported APIs which still impacts legacy+alpha API
surfaces leaking all the base classes into the API surface.

This change introduces a new pattern, where a new purely internal
interface is implemented, `SharedKernelFactory`, and passed to
`makeSharedObjectKind` which handles everything else. The
`SharedKernelFactory` exposes the impelementation of the SharedObject
policy as the new `SharedKernel` interface. This object is separate from
the app facing view, making it easier to expose minimal APIs (Like
SharedTree's ITree) without leaking in APIs from SharedObject.

While currently this pattern can't support an app facing API with a
property which would collide with a property of SharedObject, a
limitation of all current SharedObjects, it sets up a approach which
could decouple the app facing APi from SharedObject entirely in the
future if desired, both making changes to SharedObject and to the app
facing API surfaces easier.

In addition to just being a cleaner interface to code SharedObjects
against which enable future flexibility, it also is key in enabling the
migration approach taken by
https://github.com/microsoft/FluidFramework/pull/23729 as it provides a
common abstraction that supported SharedObjects can implement
(SharedKernel) that is possible to migrate while preserving the object
identity of the SharedObject which wraps it; This scenario is the main
motivation for doing this now.

The approach of splitting out the core implementation logic for a DDS
was inspired by SharedMap which did this with its SharedMapKernel. This
didn't end up getting used to implement SharedDirectory, but the
approach seems effective and could have facilatated such use.

This pattern has been successfully applied to tree and is fully hooked
up to tree in this change.

This work is split out from
https://github.com/microsoft/FluidFramework/pull/23729 and has some
forward looking features to enable supporting SharedMap, like `thisWrap`
which is required to implement SharedMap's "set" function which returns
`this`.