Fuzz: Reduce change cascades due to random call pattern changes (#24331)
A common problem in our fuzz test is that a small change in a reducer
which affects the calling pattern on our random generator have a
cascading effect on all of the tests that are generated which can cause
the results for a test seed to shift and invalidate skipped seeds, or
change coverage.

This change attempts to add some stability to the random. Basically, we
give each generator/reducer instance its own seeded random. This won't
solve all possibilities for cascades but reduces cascades in the case
where the change to the random calling pattern is limited to a generator
or reducer. It can't wholly prevent cascades as the new call patterns
can change the state, so if future reducers/generators depend on
previous state.

We will also persist the op seeds on the ops, so they will also be
reused in replay. This shouldn't change much, as reducers shouldn't use
random, but if they do, this will ensure they behave the same on replay
as they did in the original test, which also adds a bit more stability.

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>