DDS Fuzz: support numbers or iterable for all seed based test options (#24596)
This is really about ease of use. Supporting the same types of all
options makes it easy to move between then as I run and debug the tests
without requiring any re-formatting.

For instance, i can now change a `.skip` to a `.only` to rerun the
skipped tests, and then i can change that to `.reply` to run the
minimized forms.