build: Update api-extractor to 7.52.8 (#24599)
## Description

Updates to version 7.52.8 of `@microsoft/api-extractor` to fix a
regression introduced in 7.52.5 that caused running it without the
`--local` flag (like our pipelines) to not fail when it should have
(e.g. if the api reports were out of date, or TSDoc links pointed to
non-existent APIs).

Also fixes some issues with api-reports that slipped in while the
regression was in place (see
https://github.com/microsoft/FluidFramework/pull/24599/commits/78bbaab9585b2ea73b342eda2ba2ba1fbc138df4).