Add `--fgrep @ExecutionTime` to perf benchmarks scripts (#24605)
## Description

Many npm scripts to run execution-time benchmarks across the repo were
missing the `--fgrep @ExecutionTime` flag to limit themselves to running
only the correct benchmarks. E.g. the tree package was running memory
tests as if they were execution-time benchmarks. By coincidence, when
using the previous version of the benchmark tool (0.50), running the
execution-time benchmarks in CI would run into a weird issue that seems
to be mocha crashing silently but terminating with an exit code of 0. So
the pipeline step succeeded even when not all benchmarks were running.
With the upgrade to benchmark 0.51 the crash doesn't happen anymore, but
that revealed the fact that running the execution-time benchmarks for
tree timed out after 60 minutes, and investigating that revealed the
problem with the missing `--fgrep @ExecutionTime` flag.

`test:benchmark:report` is the script we use in the pipelines
specifically for execution-time benchmarks, so all those were updated.
Some packages also have `perf`, `bench`, and other variants, that I
believe are almost certainly intended for the execution-time benchmarks
as well, so I updated those too. Only the `test:memory-profiling:report`
script uses the `--fgrep @MemoryUsage` flag, and only
`test:customBenchmarks` uses `@CustomBenchmark`, so those weren't
touched.