Simplify, document and improve type safety and naming in hydration logic (#24624)
## Description

The hydration logic erroneously calls TreeNodes proxies. Not only are
many tree nodes not implemented using proxies, but many other things
are, so use the more specific and correct term, TreeNode (or node when
not neeing to be specific).

Add documentation to many undocumented things.

Fix incorrect and out of date documentation.

Deduplicate the logic for single items and arrays by using an array of a
single item.

Improve type safety for how the root of paths are edited.

Fixing up TreeNodeKernel.anchorProxy is out of scope: that has been left
referring to proxy and will be looked at independently.

Moving/renaming the tests has also been skipped to help ensure the tests
as is pass with these changes. They will also be moved and looked at
independently.