Re-enable skipped conflicting ops test (#24663)
## Description

This test timed out once (maybe twice though we don't have the logs at
this point) in the last few months on local-server, but upon scrutiny I
don't believe it has any race conditions and the failure was more likely
due to the CI machine's resources being overloaded. With the fixes in
#24647, that problem should be mitigated.

I also ran this test repeatedly locally by looping it and didn't see any
issues for some additional peace-of-mind.

AB#29237

Co-authored-by: Abram Sanderson <absander@microsoft.com>