refactor(api-markdown-documenter): Simplify DocumentationNode domain (#24668)
Removes the type concepts of `SingleLine` and `MultiLine` documentation
nodes. Instead, rendering of certain kinds of nodes into Markdown syntax
will fall back to HTML where necessary. This includes list nodes.

Also updates `CodeSpan` to only accept plain text children, which is the
only content allowed in that context in Markdown and is the only way it
was used by the system anyways.