fix(server): NexusSessionResult timestamp and duration bugs (#24556)
## Description

Recently, a few bugs/issues were found in the NexusSessionResult
telemetry event.

1. DurationInMs could sometimes be ~1.7Trillion, which is roughly equal
to `Date.now()` or "mx since epoch."
- It appears that in some cases, `IClient.timestamp` was possibly `0`.
We shouldn't have been relying on this anyways, since we already have
`connectedTimestamp` tracked in the service.
- There was also a suspicious `joinedTime ?? 0` which didn't seem to be
in the path for telemetry use, but I removed it completely just to make
sure.
2. Duplicate, overlapping "sessions" were possible because a client
could re-connect to a new instance within the session activity timeout
window, but on a new Nexus instance. This would allow the old instance's
timer to remain uncleared.
- A quick "make sure the session wasn't updated recently" check on timer
firing was sufficient for this bug
3. The timestamp for `NexusSessionResult` telemetry event aligned with
the _end_ of the event, rather than the _start_ of the event. All other
telemetry events' timestamps align with the start of the event.
    - A new `overrideTimestamp` method was added to the `Lumber` class.

## Breaking Changes

A few internal classes/methods have new/altered/reduced parameters.
These are internal, though, and shouldn't cause any issues.

## Reviewer Guidance

I locally made some alterations for testing this code using multiple
Nexus instances within the Docker Compose behind NGINX with Round Robin
loadbalancing. This fairly accurately simulates having multiple Nexus
instances in our production setup. I did not include this change in the
PR, as there is not a good way to enable/disable it.

## WIP

- [x] UTs