Directory: Add minimization transform (#24693)
Adds a simple minimization transform for directory fuzz tests. By
reducing the path depth it will allow the removal of things like create
subdirectory ops, which can lead to a much simpler repo, especially in
cases like local server fuzz where we might just be storing a handle in
a root dds to attach it, so don't need it in a deep sub dir.