fix(id-allocation): Schedule a flush when submitting ID Allocation op before replaying pending states (#24683)
In `ContainerRuntime`, we submit to the `Outbox` directly from
`replayPendingStates`, and so we should also call `scheduleFlush` to
ensure that op doesn't get stuck in the Outbox. Being "stuck in the
Outbox" is a problem because it's assumed that all ops in a batch come
during the same JS turn and have the same reference sequence number, and
this can easily lead to violating that invariant.

That one-line change is accompanied by some refactoring so we don't have
to consider Immediate mode at that callsite (`scheduleFlush` handles it
now).