Revert `Deprecate containerPackageInfo parameter in createOdspCreateContainerRequest` (#24481)
## Description

When communicating with partners about work of [issue
#23882](https://github.com/microsoft/FluidFramework/issues/23882), I
realized that the code that was meant to be deleted will actually still
be used, so there no longer is any need for the deprecation and removal.

This PR is meant to be a revert for this one [PR
#23919](https://github.com/microsoft/FluidFramework/pull/23919/files)

## Breaking Changes

This will close [issue
#23882](https://github.com/microsoft/FluidFramework/issues/23882)
without actually deleting any code.

## Reviewer Guidance

Let me know if you have any questions related to the revert or the code
in general.

Fixes:
[AB#31049](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31049)