Add telemetry for connection attempt in ODSP for debugging purposes (#24706)
## Description

Add telemetry for connection attempt in ODSP for debugging purposes.
Loop web service has 1 min timeout for container to get connected before
they timeout. So, we need to record this telemetry before we initiate
the connection as it happens that we may not get the later telemetry due
to timeout.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>