Fuzz Minimization: Fix off by one error (#24732)
This pull request includes a small but significant change to the
`FuzzTestMinimizer` class in the `stochastic-test-utils` package. The
change adjusts the loop condition in the `minification.ts` file to
ensure the loop iterates over all operations, including the first one.

*
[`packages/test/stochastic-test-utils/src/minification.ts`](diffhunk://#diff-9c112b4f8da82c2073195da15003b53b84b8cebc2edbe8a217a0d6a38fc40518L94-R94):
Changed the loop condition in the `FuzzTestMinimizer` class from `idx >
0` to `idx >= 0`, ensuring that the first operation in the list is also
considered during the loop.