Tree oldestCompatibleClient plumbing (#24237)
## Description

Plumb `oldestCompatibleClient: FluidClientVersion` through the tree code
from SharedObject configuration into the `Summarizable`s.

This does not actually make any encoding changes based on the version
specified, but future changes will do so. Therefor any code specifying
the `oldestCompatibleClient` version should ensure it does not pass in
too high of a version.

A few related types have been marked `@input`: since they are already
`@alpha` this does not change anything usage contract wise (they could
already change arbitrarily): it is just clarifying what the support
contract will be if/when they become public/stable and what the intended
usage patterns are.

Adds and tests, but does not yet use
`ClientVersionDispatchingCodecBuilder` which will be able to finish
plumbing the version through to select formats.

Updates the error when an unknown version is encountered to be a usage
error instead of an assert since this will be able to be hit by old
client legitimately once newer one start encoding in newer formats.